Redistribute util.py into surgery, features, arrays, and optional modules
Closes #49. util.py had become a dumping ground; its contents now live in
modules named for what they do:

- surgery.py: elementwise registry + shift_loc/slice_time/prefix_condition
  dispatchers and MVN registrations (kept together so import-time
  registration side effects always accompany the generics)
- features.py: fourier_features, periodic_repeat
- arrays.py: zero_data_like, concat_future
- optional.py: require, _api_canary

All imports rewired (package, tests, README, AGENTS.md), tests/test_util.py
split into per-module test files, great-docs reference restructured into one
section per module, and the five example notebooks re-executed in place with
the new import paths.
