+++ numpyro_forecast/exceptions.py
@@ -131,9 +131,9 @@ class KernelConfigError(NumpyroForecastError, ValueError):
 class MVNLayoutError(NumpyroForecastError, NotImplementedError):
     """A ``MultivariateNormal`` layout is unsupported for time-axis surgery.
 
-    Raised by :func:`~numpyro_forecast.util.shift_loc`,
-    :func:`~numpyro_forecast.util.slice_time`, and
-    :func:`~numpyro_forecast.util.prefix_condition` on MVN noise whose
+    Raised by :func:`~numpyro_forecast.surgery.shift_loc`,
+    :func:`~numpyro_forecast.surgery.slice_time`, and
+    :func:`~numpyro_forecast.surgery.prefix_condition` on MVN noise whose
     ``loc``/``covariance_matrix`` shapes do not match the supported
     time-leading layout.
     """
