+++ tests/test_examples.py
@@ -9,9 +9,9 @@ from jax import Array, random
 
 from numpyro_forecast.datasets import bart_available, load_bart_hierarchical, load_bart_weekly
 from numpyro_forecast.evaluate import eval_crps
+from numpyro_forecast.features import fourier_features
 from numpyro_forecast.forecaster import Forecaster, _BaseForecaster
 from numpyro_forecast.typing import ForecastModel
-from numpyro_forecast.util import fourier_features
 
 MakeForecaster = Callable[..., _BaseForecaster]
